requirements.txt
pyjwt
requests
zhipuai
python-dotenv